Resolve Compiler Warnings (#64)

+++ Source/Core/Writers/ERS_SceneWriter/ERS_SceneWriter.cpp
@@ -165,7 +165,7 @@ std::string SceneWriter::ProcessScene(std::shared_ptr<ERS_STRUCT_Scene> InputSce
     }
 
     //---- Write Spot Lights ----//
-    for (int i = 0; i < InputScene->SpotLights.size(); i++) {
+    for (int i = 0; (long)i < (long)InputScene->SpotLights.size(); i++) {
 
         Output << YAML::Key << AssetIndex;
         Output << YAML::BeginMap;
